requirements.txt
amqp==5.1.0
asgiref==3.4.1
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
boto3==1.21.4
botocore==1.24.4
cached-property==1.5.2
cachetools==5.0.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor-5 @ git+https://github.com/ycy-tw/django-ckeditor-5.git@7d5a1fdf5b250b635570d97522844f4f5cd7676d
django-js-asset==1.2.2
django-storages==1.12.3
django-taggit==1.5.1
google-analytics-data==0.11.2
google-api-core==2.7.2
google-auth==2.6.4
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
jmespath==0.10.0
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
zipp==3.8.0
